feat(lucy_config_pipeline): add firmware build phase and refactor package layout
