+++ lucy_config_pipeline/lucy_config_pipeline/pipeline_build.py
@@ -1,0 +1,5 @@
+"""Backward-compatible import wrapper."""
+
+from .pipeline.build import run_build_phase
+
+__all__ = ["run_build_phase"]
