+++ lucy_config_pipeline/lucy_config_pipeline/pipeline_models.py
@@ -1,0 +1,5 @@
+"""Backward-compatible import wrapper."""
+
+from .pipeline.models import FirmwarePaths, PipelinePaths
+
+__all__ = ["PipelinePaths", "FirmwarePaths"]
