+++ lucy_config_pipeline/lucy_config_pipeline/pipeline_selection.py
@@ -1,0 +1,15 @@
+"""Backward-compatible import wrapper."""
+
+from .pipeline.selection import (
+    board_build_plan,
+    resolve_firmware_paths,
+    resolve_mapping_input,
+    select_boards_to_process,
+)
+
+__all__ = [
+    "resolve_mapping_input",
+    "select_boards_to_process",
+    "resolve_firmware_paths",
+    "board_build_plan",
+]
